Index data under the same index name
Before this change the index name was based on the database filename and
the document type was based on the table name. Now, the index name is
'sqlite' by default and the document type depends on both the database
filename and the table name. This makes more sense since everything is
under the same index.